Allow downloading a fixture in a converted format
